requirements.txt
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.0.0
openpyxl>=3.1.0
joblib>=1.3.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
firebase-admin>=6.2.0
aiohttp>=3.9.0
pydantic>=2.0.0
scholarly>=1.7.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

